Move ComponentDescriptor validation out of BindingGraphValidator and into a new ComponentDescriptorValidator.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=200455229
